backend/requirements.txt
celery==5.3.4
redis==5.0.1
pymongo==4.6.1
python-dotenv==1.0.0
requests==2.31.0
flask==3.0.0 